Commands MK1
Updated POST request things from program 1 - 2. When the message SEND_COUNTRY is typed into the textbox on the form and sent to program 2, it will recognise that. Next it will look into the class ReturnCountry.cs. This retrieves the country ID and name. This then gets written to the form.

ID isnt currently output, might be an easy fix, havent even bothered to look. Taking a break.