fix: recovery mode processes
An in depth re work of the recovery mode and batch processing. It now has been fully implemented and tested, transactions and bookings can be saved for later processing in the event of a failure. recovery mode is able to retrieve and processes batches. some general reworking of code for optimisation as well.